add packing in if statement

+++ main_process.js
@@ -15,34 +15,36 @@ app.on('ready', () => {
   //Adding Menu Bar
   const menu = Menu.buildFromTemplate(createMenuBar(mainWindow, ResetDir, OpenDir));
   Menu.setApplicationMenu(menu);
-  const initialStartFlagFilePath = path.join(__dirname, 'electronUserData', 'initialStartup.txt');
-  if (!fs.existsSync(initialStartFlagFilePath)) {
-    let pathToExecutable;
-    if (process.platform === 'darwin') {
-      pathToExecutable = path.join(__dirname, '..', '..', 'MacOS', 'bundle-bee');
-    } else if (process.platform === 'win32') {
-      pathToExecutable = path.join(__dirname, '..', '..', 'MacOS', 'bundle-bee');
-    }
-    fs.writeFile(
-      initialStartFlagFilePath,
-      `command to run executable:
+  if (process.ENV === 'package') {
+    const initialStartFlagFilePath = path.join(__dirname, 'electronUserData', 'initialStartup.txt');
+    if (!fs.existsSync(initialStartFlagFilePath)) {
+      let pathToExecutable;
+      if (process.platform === 'darwin') {
+        pathToExecutable = path.join(__dirname, '..', '..', 'MacOS', 'bundle-bee');
+      } else if (process.platform === 'win32') {
+        pathToExecutable = path.join(__dirname, '..', '..', 'MacOS', 'bundle-bee');
+      }
+      fs.writeFile(
+        initialStartFlagFilePath,
+        `command to run executable:
     open ${pathToExecutable}
     dirname: ${__dirname}
     filename: ${__filename}
     initalstartflagfilepath: ${initialStartFlagFilePath}
     `
-    );
-    exec(`open ${pathToExecutable}`, err => {
+      );
+      exec(`open ${pathToExecutable}`, err => {
+        app.exit(0);
+      });
+    } else {
+      fs.unlink(initialStartFlagFilePath, err => {
+        if (err) console.log(err);
+      });
+    }
+    mainWindow.on('close', () => {
       app.exit(0);
     });
-  } else {
-    fs.unlink(initialStartFlagFilePath, err => {
-      if (err) console.log(err);
-    });
   }
-  mainWindow.on('close', () => {
-    app.exit(0);
-  });
 });
 
 ipcMain.on('ondragstart', (event, filePath) => {
@@ -178,4 +180,4 @@ function ResetDir() {
   app.exit(0);
 }
 
-function OpenDir(build) {}
+function OpenDir(build) { }
